Refactoring Gameboard's data management into a class of its own, holding states and data without user interactions or knowing about the View
