+++ client/src/app/gameboard/board-cloner.ts
@@ -1,0 +1,10 @@
+import { Square, FileEnum } from './square';
+import { Move } from './move';
+
+export default class BoardCloner {
+  static cloneBoard(currentBoard: Square[][]): Square[][] {
+    const result: Square[][] = [];
+
+    return result;
+  }
+}
