+++ client/src/app/gameboard/gameboard.component.html
@@ -1,10 +1,10 @@
 <div class="text-center">
-  <h4>current turn: {{ currTurn }}</h4>
-  <h5>{{ whiteKingChecked || blackKingChecked ? 'checked' : '' }}</h5>
+  <h4>current turn: {{ gb.currTurn }}</h4>
+  <h5>{{ gb.whiteKingChecked || gb.blackKingChecked ? 'checked' : '' }}</h5>
   <div class="board d-table mx-auto">
     <div
       *ngFor="
-        let r of board.slice().reverse();
+        let r of gb.board.slice().reverse();
         let r_index = index;
         let r_odd = odd;
         let r_even = even
@@ -28,23 +28,25 @@
       >
         <!-- activate square if a piece is trying to move -->
         <div
-          *ngIf="moving && currMovesMap.has('' + s.file + s.rank)"
+          *ngIf="gb.moving && gb.currMovesMap.has('' + s.file + s.rank)"
           class="move"
         ></div>
         <div
           class="current"
           *ngIf="
-            moving && currSquare.file === s.file && currSquare.rank === s.rank
+            gb.moving &&
+            gb.currSquare.file === s.file &&
+            gb.currSquare.rank === s.rank
           "
         ></div>
         <div
           class="offensive"
           *ngIf="
             isAdmin &&
-            ((currTurn == 'white' &&
-              attackMovesMap.black.has('' + s.file + s.rank)) ||
-              (currTurn == 'black' &&
-                attackMovesMap.white.has('' + s.file + s.rank)))
+            ((gb.currTurn == 'white' &&
+              gb.attackMovesMap.black.has('' + s.file + s.rank)) ||
+              (gb.currTurn == 'black' &&
+                gb.attackMovesMap.white.has('' + s.file + s.rank)))
           "
         ></div>
         <i
@@ -64,7 +66,7 @@
   <h4>captured pieces:</h4>
   <p>
     <i
-      *ngFor="let piece of capturedPieces"
+      *ngFor="let piece of gb.capturedPieces"
       class="fas fa-chess-{{ piece?.name }}"
       [ngClass]="{
         white: piece?.color == 'white',
